### 2018-04-09 #669 @whedon
[PRE REVIEW]: BioSigKit: A Matlab Toolbox and Interface for Analysis of BioSignals
**Submitting author:** @hooman650 (<a href="http://orcid.org/0000-0003-3577-2687">Hooman Sedghamiz</a>)
**Repository:** <a href="https://github.com/hooman650/BioSigKit" target ="_blank">https://github.com/hooman650/BioSigKit</a>
**Version:** V1.0.0
