### 2018-04-09 da69d72
Fixed URLs for citations
