+++ paper/paper.bib
@@ -63,15 +63,17 @@ year = {2012}
 
 @article{sedghamiz2014complete,
   title={Complete Pan-Tompkins Implementation ECG QRS Detector},
-  author={Sedghamiz, H},
-  journal={Matlab Central: Community Profile. Available online at: http://www. mathworks. com/matlabcentral/profile/authors/2510422-hooman-sedghamiz},
+  author={H. Sedghamiz},
+  journal={Matlab Central: Community Profile. Available online.},
+  url = {http://www. mathworks. com/matlabcentral/profile/authors/2510422-hooman-sedghamiz},
   year={2014}
 }
 
 @article{sedghamiz2013online,
   title={An online algorithm for r, s and t wave detection},
-  author={Sedghamiz, Hooman},
-  journal={Matlab Central: Community Profile. Available online at: https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector?s_tid=prof_contriblnk},
+  author={H. Sedghamiz},
+  journal={Matlab Central: Community Profile. Available online.},
+  url = {https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector?s_tid=prof_contriblnk},
   year={2013}
 }
 
