+++ paper/paper.bib
@@ -61,20 +61,20 @@ volume = {5},
 year = {2012}
 }
 
-@article{sedghamiz2014complete,
-  title={Complete Pan-Tompkins Implementation ECG QRS Detector},
-  author={Sedghamiz, Hooman},
-  journal={Matlab Central: Community Profile; Available online},
-  url = {http://www. mathworks. com/matlabcentral/profile/authors/2510422-hooman-sedghamiz},
-  year={2014}
-}
-
 @article{sedghamiz2013online,
-  title={An online algorithm for r, s and t wave detection},
-  author={Sedghamiz, Hooman},
-  journal={Matlab Central: Community Profile; Available online},
+  title={{An online algorithm for r, s and t wave detection}},
+  author={Sedghamiz, H},
+  journal={Matlab Central Community Profile Available online},
   url = {https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector?s_tid=prof_contriblnk},
   year={2013}
+}
+
+@article{sedghamiz2014complete,
+  title={{Complete Pan-Tompkins Implementation ECG QRS Detector}},
+  author={Sedghamiz, H},
+  journal={Matlab Central Community Profile Available online},
+  url = {http://www. mathworks. com/matlabcentral/profile/authors/2510422-hooman-sedghamiz},
+  year={2014}
 }
 
 @INPROCEEDINGS{7391510, 
