+++ paper/paper.bib
@@ -82,10 +82,8 @@ author={H. Sedghamiz and D. Santonocito},
 booktitle={2015 E-Health and Bioengineering Conference (EHB)}, 
 title={Unsupervised detection and classification of motor unit action potentials in intramuscular electromyography signals}, 
 year={2015}, 
-volume={}, 
-number={}, 
 pages={1-6}, 
 keywords={electromyography;medical signal detection;pattern clustering;principal component analysis;signal classification;unsupervised learning;MTEO based analysis method;MUAP detection;PCA;computational speed;correlation computation load;diseased muscles;fiducial point detection;healthy muscles;intramuscular EMG recordings;intramuscular electromyography signals;label matching techniques;motor unit action potential classification;motor unit action potential clustering;motor unit action potential detection;multiresolution Teager energy operator;posterior cricoarytenoid anterior;principal component analysis;single-channel intramuscular electromyography;template matching techniques;tibiliasis anterior;unsupervised clustering;unsupervised detection;Algorithm design and analysis;Band-pass filters;Clustering algorithms;Electric potential;Electromyography;Noise measurement;Principal component analysis;decomposition;electromyography;template matching;unsupervised classification}, 
 doi={10.1109/EHB.2015.7391510}, 
-ISSN={}, 
-month={Nov},}
+month={Nov}
+}
