+++ paper/paper.bib
@@ -64,16 +64,16 @@ year = {2012}
 @article{sedghamiz2013online,
   title={{An online algorithm for r, s and t wave detection}},
   author={Sedghamiz, H},
-  journal={Matlab Central Community Profile Available online},
+  journal={Matlab Central: Community Profile. Available online at},
   url = {https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector?s_tid=prof_contriblnk},
   year={2013}
 }
 
 @article{sedghamiz2014complete,
-  title={{Complete Pan-Tompkins Implementation ECG QRS Detector}},
+  title={Complete Pan-Tompkins Implementation ECG QRS Detector},
   author={Sedghamiz, H},
-  journal={Matlab Central Community Profile Available online},
-  url = {http://www. mathworks. com/matlabcentral/profile/authors/2510422-hooman-sedghamiz},
+  journal={Matlab Central: Community Profile. Available online at},
+  url={https://www.mathworks.com/matlabcentral/fileexchange/45840-complete-pan-tompkins-implementation-ecg-qrs-detector},
   year={2014}
 }
 
