+++ paper/paper.bib
@@ -72,8 +72,7 @@ year = {2012}
 @article{sedghamiz2014complete,
   title={Complete Pan-Tompkins Implementation ECG QRS Detector},
   author={Sedghamiz, H},
-  journal={Matlab Central: Community Profile. Available online at},
-  url={https://www.mathworks.com/matlabcentral/fileexchange/45840-complete-pan-tompkins-implementation-ecg-qrs-detector},
+  journal={Matlab Central: Community Profile. Available online at: http://www. mathworks. com/matlabcentral/profile/authors/2510422-hooman-sedghamiz},
   year={2014}
 }
 
