+++ paper/paper.bib
@@ -18,7 +18,7 @@ year = {2002}
 abstract = {We have developed a real-time algorithm for detection of the QRS complexes of ECG signals. It reliably recognizes QRS complexes based upon digital analyses of slope, amplitude, and width. A special digital bandpass filter reduces false detections caused by the various types of interference present in ECG signals. This filtering permits use of low thresholds, thereby increasing detection sensitivity. The algorithm automatically adjusts thresholds and parameters periodically to adapt to such ECG changes as QRS morphology and heart rate. For the standard 24 h MIT/BIH arrhythmia database, this algorithm correctly detects 99.3 percent of the QRS complexes.},
 author = {Pan, Jiapu and Tompkins, Willis J.},
 doi = {10.1109/TBME.1985.325532},
-isbn = {0018-9294 VO  - BME-32},
+isbn = {0018-9294VO-BME-32},
 issn = {0018-9294},
 journal = {IEEE Transactions on Biomedical Engineering},
 number = {3},
@@ -64,15 +64,16 @@ year = {2012}
 @article{sedghamiz2013online,
   title={{An online algorithm for r, s and t wave detection}},
   author={Sedghamiz, H},
-  journal={Matlab Central: Community Profile. Available online at},
+  journal={Matlab Central Community Profile Available online},
   url = {https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector?s_tid=prof_contriblnk},
   year={2013}
 }
 
 @article{sedghamiz2014complete,
-  title={Complete Pan-Tompkins Implementation ECG QRS Detector},
+  title={{Complete Pan-Tompkins Implementation ECG QRS Detector}},
   author={Sedghamiz, H},
-  journal={Matlab Central: Community Profile. Available online at: http://www. mathworks. com/matlabcentral/profile/authors/2510422-hooman-sedghamiz},
+  journal={Matlab Central Community Profile Available online},
+  url = {https://www.mathworks.com/matlabcentral/fileexchange/45840-complete-pan-tompkins-implementation-ecg-qrs-detector},
   year={2014}
 }
 
