+++ paper/paper.bib
@@ -62,19 +62,21 @@ year = {2012}
 }
 
 @article{sedghamiz2013online,
-  title={{An online algorithm for r, s and t wave detection}},
-  author={Sedghamiz, H},
-  journal={Matlab Central Community Profile Available online},
+  title = {{An online algorithm for r, s and t wave detection}},
+  author = {Sedghamiz, H},
+  doi = {10.13140/RG.2.2.23774.64328},
+  journal = {Matlab Central Community Profile Available online},
   url = {https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector?s_tid=prof_contriblnk},
-  year={2013}
+  year = {2013}
 }
 
-@article{sedghamiz2014complete,
-  title={{Complete Pan-Tompkins Implementation ECG QRS Detector}},
-  author={Sedghamiz, H},
-  journal={Matlab Central Community Profile Available online},
+@article{sedghamiz2014completed,
+  title = {{Complete Pan-Tompkins Implementation ECG QRS Detector}},
+  doi = {10.13140/RG.2.2.14202.59841},
+  author = {Sedghamiz, H},
+  journal = {Matlab Central Community Profile Available online},
   url = {https://www.mathworks.com/matlabcentral/fileexchange/45840-complete-pan-tompkins-implementation-ecg-qrs-detector},
-  year={2014}
+  year = {2014}
 }
 
 @INPROCEEDINGS{7391510, 
