### 2018-04-09 3f46b5d
Update paper.md
