### 2018-04-11 #669 @Kevin-Mattheus-Moerman
@cMadan @keevindoherty @germangh @marianux would you be interested in reviewing this [JOSS](http://joss.theoj.org/) submission? Here is the [article](https://github.com/openjournals/joss-papers/blob/joss.00669/joss.00669/10.21105.joss.00669.pdf) and this is the [project repo](https://github.com/hooman650/BioSigKit). The review focuses largely on the software, see our [review guidelines](http://joss.theoj.org/about#reviewer_guidelines).
