### 2018-04-11 #669 @arfon
> Hi, @arfon , whedon was not able to compile my .bib file properly, I attempted several times and the .bib file looks okay but the compiled pdf does not. Apologies if I was not supposed to rebuild the pdf...

@hooman650 - could you describe the problem? The PDF looks OK to me.
