### 2018-04-11 #669 @marianux
Yes, happy to review it. When is the deadline ?

Best,
m.

2018-04-11 13:28 GMT-03:00 Kevin Mattheus Moerman <notifications@github.com>
:

> @cMadan <https://github.com/cMadan> @keevindoherty
> <https://github.com/keevindoherty> @germangh <https://github.com/germangh>
> @marianux <https://github.com/marianux> would you be interested in
> reviewing this JOSS <http://joss.theoj.org/> submission? Here is the
> article
> <https://github.com/openjournals/joss-papers/blob/joss.00669/joss.00669/10.21105.joss.00669.pdf>
> and this is the project repo <https://github.com/hooman650/BioSigKit>.
> The review focuses largely on the software, see our review guidelines
> <http://joss.theoj.org/about#reviewer_guidelines>.
>
