### 2018-04-11 #669 @arfon
> @arfon - the name of the author is missing in one of the references (the one prior to the last-"Complete PAn-Tompkins implemetation" )...

This is very strange. I have literally no idea what's going on here. I tried debugging this locally but couldn't easily fix. I suggest we proceed with the submission and revisit this later.
