### 2018-04-11 #671 @Kevin-Mattheus-Moerman
@marianux you asked about a deadline. Typical review times range from days or weeks to months for larger projects with many issues. There is no official deadline but we obviously encourage a smooth and reasonably fast review process.
