### 2018-04-11 #669 @cMadan
@hooman650, @arfon - Isn't the reference/bib 'issue' just the replacement of a repeated author across references with the dash?
(E.g., https://tex.stackexchange.com/questions/29381/is-it-normal-for-bibtex-to-replace-similar-author-names-with)

Is there something that I'm missing?
