### 2018-04-12 #671 @hooman650
@Kevin-Mattheus-Moerman, within the list of JOSS reviewers, @jordigh , @nicoguaro, @sealhuang @mosteo, @nirum ;  and out of that list  I can think of @fernandoandreotti , also @marianux has done his PhD on ECG analysis and has also made a very nice toolbox so he might be able to suggest someone else from his circle?
