### 2018-04-13 #671 @Kevin-Mattheus-Moerman
Thanks @marianux @fernandoandreotti for agreeing to review this submission.

Review is centered largely on the software and the below paper. You use the above tick boxes to guide the review process. You can comment and provide feedback in this review issue. You are also encouraged to open issues on the [project repository](https://github.com/hooman650/BioSigKit) and link to them here.

Here are the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines). Let me know if you have any questions.

Let the reviewing begin! :rocket: :robot:
