+++ README.md
@@ -1,5 +1,5 @@
 # BioSigKit
-BioSigKit is a set of useful signal processing tools in Matlab that are either developed by me personally or others in different fields of biosignal processing. BioSigKit is a wrapper with a simple visual interface that gathers this tools under a simple easy to use platform. 
+BioSigKit is a set of useful signal processing tools in Matlab that are either developed by me personally or others in different fields of biosignal processing. BioSigKit is a wrapper with a simple visual interface that gathers this tools under a simple easy to use platform. The ultimate goal of BioSigKit is not to be only used for ECG processing, it aims to be helpful in analysis of several different physiological signals such as EMG, ACC and EDR as well. For instance the last subroutine which is a general peak detector can already be used in many different signals even with low frequencies (Please see [the paper by Scholkmann](http://www.mdpi.com/1999-4893/5/4/588). While its current version still does not offer more algorithms for processing of the other physiological signals, at least several more subroutines are in pipeline that will be able to do so in the near future. For instance,extracting Fetal ECG from maternal ECG using projection techniques, computing ECG derived respiration (EDR) or MUAP detection in EMG signals ([EMG toolbox (offered by the same author)](https://www.mathworks.com/matlabcentral/fileexchange/59344-toolbox-for-unsupervised-classification-of-muaps-and-action-potentials-in-emg)). 
 
 # QRS Detection Algorithms offered by BioSigKit
 BioSigKit provides a set of subroutines implementing the six following QRS detection algorithms:
