+++ README.md
@@ -19,7 +19,13 @@ QRS Multilevel Teager Energy Operator (MTEO) [[H. Sedghamiz, 2016](http://ieeexp
 Automatic Multiscale-based Peak Detection [[Scholkmann2012](http://www.mdpi.com/1999-4893/5/4/588)]: This method is a more general peak detection. However, according to the study by Scholkmann et al. [[Scholkmann, 2012](http://www.mdpi.com/1999-4893/5/4/588)], it showed a high performance for the beat detection as well. Therefore, it is implemented as one of the subroutines in BioSigKit (subroutine name : ```BioSigKit.AMPD_PAlg()```).
 
 # Getting Started and Installation
-To install BioSigKit simply download the repository.
+To install BioSigKit simply:
+1. Download the repository.
+
+2. Unzip the downloaded package and simply run "RunBioSigKit.m".
+
+3. See the instrcutions bellow for further details: 
+
 BioSigKit might be used either from the command line or its Graphical User Interface (GUI). 
 ## Command Line : 
 
