### 2018-05-01 #671 @fernandoandreotti
Dear Hooman,

first of all great work implementing all these 6 QRS detectors in an object oriented way and including a GUI. The examples ran smoothly.

My concern regarding the toolbox is the fact that despite as it being advertised a toolbox for biosignals, the only implemented routines are QRS detectors. Rather than creating a new toolbox from scratch, I wonder why not add those to other toolboxes such as [WFDB](https://github.com/ikarosilva/wfdb-app-toolbox) or the [ecg-kit](http://marianux.github.io/ecg-kit/)?

Best regards,

Fernando
