### 2018-05-02 #671 @hooman650
Thank you @fernandoandreotti and @marianux for taking the time and going over the BioSigKit.

Your comments are very accurate. I understand that currently only documented functions in BioKitSig are offering QRS detectors with the exception of the last algorithm which is a general peak detector. And this might seem like reinventing the wheel beside that a few algorithms are new.

Please note that the goal of BioKitSig is to be a general biosignal processing toolbox. There are at least 5 other subroutines for processing Accelerometer signals, Fetal ECG extraction with projection methods, nonlinear projection filtering for ECG processing and ECG derived respiration computation and activity detection in EMG that are implemented but since they still do not have documentation are not included in the current version. If it is okay I will update the BioSigKit repository today adding these subroutines with documentation that would make this submission distinct from WFDB and ecg-kit.

Warm regards,
Hooman
