### 2018-05-21 fa2a477
Additional subroutines
Added many more subroutines :+1:
%% ==== List of all subroutines for ACC, EMG and etc processing ========
% (7) Activity Detection Hilbert:     obj.Env_hilbert
% (8) Comp Mobility and Complexity:   obj.ComputeHjorthP
% (9) Posture detection 3 Chann ACC:  obj.ACC_Act
% (10)PsuedoCorr template matching :  obj.TemplateMatch
% (11)ECG derived respiration :       obj.EDR_comp
% (12)ACC derived respiration :       obj.ADR_comp
%% ===== General Projective, linear and nonlinear filterings ============ %%
% (13)Real-time neural PCA:           obj.neural_pca
% (14)Adaptive Filteri
[…]
