+++ Algorithms/projective.m
@@ -1,0 +1,382 @@
+function [px] = projective (x, t, d, m, r) 
+% PROJECTIVE  
+%
+% CALL: [px] = projective (x, t, d, m, r);
+%
+% INPUTS:
+%   x: the vector containing the scalar observations
+%   t: delay parameter used in embedding
+%   d: embedding dimension to consider
+%   m: dimensiona of null space
+%   r: number of nearest neighbors to consider
+%
+% OUTPUT:
+%   px: cleaned scalar observations
+%
+% copiright (c) and written by David Chelidze 4/20/2011
+% [px] = projective (foetal_ecg(:,2), 1, 50, 45, 1500);
+ 
+if nargin < 5
+    error('need five input variables')
+end
+if nargin > 5
+    error('too many input variables, needs only five')
+end
+ 
+% check data and make a raw vector of scalar data
+if min(size(x)) ~= 1
+   error('first input should be a vector of scalar data points.')
+else
+    x=x(:)'; % just the way we need it
+end
+px = x;
+ 
+n = length(x) - d*t; % maximum number of reconstructed points possible
+
+indx = 1:n; % index of points in the embedding
+y = zeros(d+1,n);
+dy = y; % initialize corrective array
+
+% reconstruct the phase space
+for i = 1:d+1
+    y(i,:) = x(indx+(i-1)*t);
+end
+
+% define the weighting matrix
+W = eye(d+1,d+1); W(1,1) = 10^3; W(d+1,d+1) = 10^3;
+
+fprintf('Partitioning data\n')
+[kd_tree, q] = kd_partition(y, 256); % partition into kd-tree
+
+fprintf('Filtering data\n')
+for k = 1:n % find the nearest neighbors and their properties
+    [nns, ~, ~] = kd_search(y(:,k), r, kd_tree, y(:,q));
+    nns = nns{1};
+    for i = 1:d
+        nns(i,:) = nns(i,:) - mean(nns(i,:)); % remove the mean
+    end
+    D = nns*nns.'/r; % find correlation matrix
+    G = W*D*W; % transform it
+    %[E,~] = eig(G); % assuming the MATLAB gives them in increasing order
+    %E = E(:,end-1:end); 
+    %E = E(1:m);
+    % if eig does not sort correctly use the following instead:
+    [E,g] = eig(G);
+    [~,srt] = sort(diag(g));
+    E = E(:,srt(1:m));
+    nns(:,1) = (nns(:,1) - mean(nns,2));
+    dy(:,k) = W\(E*E.'*W*nns(:,1)); % corrective vector
+    dy(:,k) = mean(nns,2) + dy(:,k);
+    px(i) = mean(dy(:,k));
+end
+
+% clean the time series
+indx = [0:t:d*t, n+(0:t:d*t)];
+for q = [1:d+1, d:-1:1]
+    for l = (1+indx(q)):indx(q+1)
+        cx = 0;
+        for k = 1:q % go through all of possible corrections
+        cx = cx + dy(k,l-(k-1)*t); % all possible corrections to x(l)
+        end
+        px(l) = x(l) -  cx/q; % cleaned values
+    end
+end
+
+%--------------------------------------------------------------------------            
+function [kd_tree, r] = kd_partition(y, b, c)
+% KD_PARTITION  Create a kd-tree and partitioned database for efficiently 
+%               finding the nearest neighbors to a point in a 
+%               d-dimensional space.
+%
+% USE: [kd_tree, r] = kd_partition(y, b, c);
+%
+% INPUT:
+%   y: original multivariate data (points arranged columnwise, size(y,1)=d)
+%   b: maximum number of distinct points for each bin (default is 100)
+%   c: minimum range of point cluster within a final leaf (default is 0)
+%
+% OUTPUT:
[…]
