### 2018-05-21 f258414
Update README.md
Subroutines added
