### 2018-05-21 #671 @hooman650
Dear @marianux and @fernandoandreotti , in regards to your comments and our last correspondence, I have already added 10 subroutines to BioSigKit where these are detailed in the homepage repository of the BioSigKit. These subroutines help perform several different tasks such as: ECG-derived respiration, alarm detection in EMG signals and audio signals,  Foetal ecg extraction from maternal ECG, artifact removal from ECG recordings and many more. Please also refer to the cheat-sheet.pdf in the repository for further details and the codemap on the bottom of the repository.

Thank you
Hooman
