### 2018-05-31 #671 @Kevin-Mattheus-Moerman
@fernandoandreotti can you review @hooman650 's reply to the issues you raised?
