### 2018-06-06 #671 @hooman650
Hi @fernandoandreotti and @marianux , hope the day is treating you well! I know you are very busy and I do not mean to rush you. I have been receiving a couple of emails from some users that were asking  me how they can cite the work... the subroutines of the Biosigkit are pretty popular, for example the Pan-tompking subroutine has been downloaded more than 13,000 times ([https://www.mathworks.com/matlabcentral/fileexchange/45840-complete-pan-tompkins-implementation-ecg-qrs-detector](https://www.mathworks.com/matlabcentral/fileexchange/45840-complete-pan-tompkins-implementation-ecg-qrs-detector)) ... or this one ([https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector](https://www.mathworks.com/matlabcentral/fileexchange/45404-ecg-q-r-s-wave-online-detector) ) that has more than 6000 downloads... These are indications that the tool can be very beneficial for different aspects of research...there are many more subroutines that I have included in the tool that can be helpful for the research community... I really appreciate your feedback on this so we can make it available for the research community...
Cheers
Hooman
