### 2018-06-24 #671 @hooman650
@fernandoandreotti thank you for your feedback and taking the time to review this...I would be happy to also contribute the subroutines to other tools such as WFDB-App and FECGSYN, the aim of BioSigKit is not to introduce new algorithms that have not been exactly examined but rather gather the popular ones that have been published separately under one roof (BioSigKit)... otherwise it should have been submitted to a more specialized bioemdical signal processing journal (e.g. IEEE transactions) and not JOSS. As I already mentioned a few links in my previous message, a lot of subroutines are already being extensively employed in research and downloaded (e.g. pan-tompkins implementation). I would like to thank you again for taking the time and reviewing this...
