### 2018-07-08 349c156
Update ALE_imp.m
