### 2018-07-24 #671 @Kevin-Mattheus-Moerman
@hooman650 as you are aware you have received comments from the reviewers in this issue and also [here](https://github.com/openjournals/joss-papers/issues/6). Please try to respond to the reviewers as best you can.

@fernandoandreotti and @marianux, I aimed to summarize  your points below. Please correct me and/or statements as you see fit.

- :memo:  Comments by @fernandoandreotti

- [ ] "I think that showing the performance of the proposed algorithms, as well as clarifying the general scope of the toolbox is still necessary."

- :memo: Comments by @marianux mentions:
(For some of these you've responded [here](https://github.com/openjournals/joss-papers/issues/6), however have the changes in relation to describing the difference with respect to other packages, and why your toolbox should be an independent package been added to the paper/documentation?)

- [ ] The main concern would be that there are several toolboxes that pursues the same goal. I think that the author would emphasize which are the differences with other existing approaches, or which advantages would this toolbox have.

- [ ] Work needed on documentation, installation instructions, and API documentation

- :memo: Comments by me

- [ ] Both reviewers mention something like "Maybe you should reconsider not reinventing the wheel, and collaborating with other open-source projects." To address these types of statements I feel you can work on pointing out the difference/overlap in the nature of your toolbox and the other works. Also clearly state the value of your collected/novel works. For JOSS it is not a requirement that the work is 100% novel/different. Novel collections of tools and novel implementations in different languages can be acceptable also. So please address the reviewers comments as best you can in the paper and documentation.
