+++ paper/paper.bib
@@ -89,3 +89,12 @@ keywords={electromyography;medical signal detection;pattern clustering;principal
 doi={10.1109/EHB.2015.7391510}, 
 month={Nov}
 }
+
+@INPROCEEDINGS{Soria2015, 
+author={AJ Demski and M Llamedo Soria}, 
+booktitle={Journal of Open Research Software}, 
+title={{Unsupervised detection and classification of motor unit action potentials in intramuscular electromyography signals}}, 
+year={2016}, 
+pages={4}, 
+doi={http://doi.org/10.5334/jors.86}
+}
