+++ paper/paper.bib
@@ -98,3 +98,16 @@ year={2016},
 pages={4}, 
 doi={http://doi.org/10.5334/jors.86}
 }
+
+@INPROCEEDINGS{1380953, 
+author={Dongho Han and Y. N. Rao and J. C. Principe and K. Gugel}, 
+booktitle={2004 IEEE International Joint Conference on Neural Networks (IEEE Cat. No.04CH37541)}, 
+title={Real-time PCA (principal component analysis) implementation on DSP}, 
+year={2004}, 
+volume={3}, 
+number={}, 
+pages={2159-2162 vol.3}, 
+keywords={digital signal processing chips;principal component analysis;assembly language;floating point arithmetic;optimisation;online temporal PCA algorithm;principal component analysis;floating point DSP;digital signal processor;statistical technique;learning algorithm;real time applications;assembly language;time varying input;spacial signals;optimization;principal component estimation;tracking;Digital signal processing;Principal component analysis;Signal processing algorithms;Covariance matrix;Convergence;Application software;Digital signal processing chips;Symmetric matrices;Delay lines;Neural engineering}, 
+doi={10.1109/IJCNN.2004.1380953}, 
+ISSN={1098-7576}, 
+month={July},}
