+++ paper/paper.bib
@@ -111,3 +111,12 @@ keywords={digital signal processing chips;principal component analysis;assembly 
 doi={10.1109/IJCNN.2004.1380953}, 
 ISSN={1098-7576}, 
 month={July},}
+
+@article{sedghamiz2014completed,
+  title = {{Nonlinear noise reduction for electrocardiograms}},
+  doi = {https://doi.org/10.1063/1.166148},
+  author = {Schreiber, T and Kaplan, T},
+  journal = {Chaos},
+  url = {https://aip.scitation.org/doi/pdf/10.1063/1.166148},
+  year = {1996}
+}
