### 2018-08-20 0353583
Update paper.bib
