+++ paper/paper.bib
@@ -112,7 +112,7 @@ doi={10.1109/IJCNN.2004.1380953},
 ISSN={1098-7576}, 
 month={July},}
 
-@article{sedghamiz2014completed,
+@article{Schreiber1996,
   title = {{Nonlinear noise reduction for electrocardiograms}},
   doi = {https://doi.org/10.1063/1.166148},
   author = {Schreiber, T and Kaplan, T},
@@ -120,3 +120,4 @@ month={July},}
   url = {https://aip.scitation.org/doi/pdf/10.1063/1.166148},
   year = {1996}
 }
+
