### 2018-08-20 f67e5c4
Update paper.md
