+++ paper/paper.md
@@ -18,32 +18,61 @@ bibliography: paper.bib
 ---
 
 # Summary
-BioSigKit is a set of Matlab (The MathWorks Inc., Natick, USA) tools for analysis and visualization of bio-signals, specifically Electrocardiogram (ECG) recordings. Matlab is a widely used programming language among researchers thanks to its simple and flexible syntax. Biomedical signal processing is one of the main areas that has been benefiting from Matlab for research and rapid prototyping. One of the most widely used and studied bio-signals is ECG. ECG is a signal reflecting the rotation of the electrical heart axis in respect to the electrode position in space. Therefore, ECG provides an invaluable insight into the cardiovascular and heart functionality. The beats in an ECG recording are identified with the help of an automated QRS detection algorithm and their morphology is studied in order to better understand the underlying cardiovascular function. Many of the most popular QRS detection algorithms [@Lee2002;@Afonso1999;@Pan1985;@Scholkmann2012] are only implemented in C or other similar higher level programming languages. BioSigKit provides a set of Matlab methods that implement 6 popular ECG QRS detectors. The object oriented implementation of BioSigKit makes it easy to update and add new algorithms to its collection. The ultimate goal of BioSigKit is to provide an easy to use interactive Matlab software that provides easy access to many standard bio-signal processing algorithms. BioSigKit currenlty provides six QRS detection algorithms detailed in the next section.
+BioSigKit is a set of Matlab (The MathWorks Inc., Natick, USA) tools for analysis and visualization of bio-signals. Matlab is a widely used programming language among researchers thanks to its simple and flexible syntax. Biomedical signal processing is one of the main areas that has been benefiting from Matlab for research and rapid prototyping. BioSigKit is a collection of signal processing tools for analysis of ECG, EEG, EMG, 3 Channel Accelerometer recordings and Bioimpedance. While there are already tools such as ECG-Kit [Soria2015] that offer specialized algorithms for ECG processing, BioSigKit is a more general purpose biosignal analysis toolbox that allows processing various biological signals. For the ECG processing, BioSigKit offers several popular algorthims that with the exception of [@Pan1985] are only available in C language. BioSigKit offers the Matlab implementation of [@Lee2002;@Afonso1999;@Scholkmann2012] for QRS detection and analysis along with several other algorithms that are detailed below. BioSigKit also provides subtourtines for activity detection in EMG recordigs, posture estimation from 3 channel Accelerometers and several adaptive filtering routines The object oriented implementation of BioSigKit makes it easy to update and add new algorithms to its collection. The ultimate goal of BioSigKit is to provide an easy to use interactive Matlab software that provides easy access to many standard bio-signal processing algorithms. BioSigKit currenlty provides six QRS detection algorithms detailed in the next section.
 
 ![Graphical User Interface of BioSigKit. The algorithm pop-up menu provides an easy way for the selection of the QRS detection algorithm. The statistics panel automatically computes mean, maximum and minimum detected intervals.](fig1.png)
 
 # BioSigKit Algorithms and Use
-BioSigKit provides a set of subroutines implementing the six following QRS detection algorithms:
+BioSigKit provides a set of subroutines for analysis of several physiological recordings detailed below.
 
-### 1. Pan and Tompkins [@Pan1985;@sedghamiz2014completed]: 
+## ECG Analysis tools:
+
+### 1.1. Pan and Tompkins [@Pan1985;@sedghamiz2014completed]: 
[…]
