### 2018-08-20 744f086
Update paper.md
