### 2018-08-20 626bbf0
Update paper.md
