### 2018-08-20 #671 @hooman650
@Kevin-Mattheus-Moerman , I have update the paper.bib file and explained the differences between the BioSigKit toolbox and other toolboxes such as ecg-kit.  The respository is updated as detailed below:
1. CheatSheet.pdf added that details all subroutines of the toolbox
2. Demo.m that runs a demo and example of all subroutines
3. paper.bib is updated with the details of the toolkit.
4. A Codemap is used in the repository that details all the subroutines for future collaboration.
5. The readme.md is also updated that details the use of the toolbox.

Hope that I have addressed the comments properly.

Have a good weekend,
Hooman
