### 2018-08-21 37c1196
Update paper.bib
