+++ paper/paper.bib
@@ -93,7 +93,7 @@ month={Nov}
 @INPROCEEDINGS{Soria2015, 
 author={AJ Demski and M Llamedo Soria}, 
 booktitle={Journal of Open Research Software}, 
-title={{Unsupervised detection and classification of motor unit action potentials in intramuscular electromyography signals}}, 
+title={{ecg-kit: a Matlab Toolbox for Cardiovascular Signal Processing}}, 
 year={2016}, 
 pages={4}, 
 doi={http://doi.org/10.5334/jors.86}
