### 2018-08-27 #671 @marianux
Hi all, my suggestion was to collaborate with another open-source project
that already did the same. IMHO the work was important, but not enough
different from others already published.

Best,
M.

El lun., 27 de ago. de 2018 a la(s) 04:23, Kevin Mattheus Moerman (
notifications@github.com) escribió:

> @fernandoandreotti <https://github.com/fernandoandreotti> @marianux
> <https://github.com/marianux> thanks again for your initial review. Could
> you please provide feedback on how the author has now dealt with your
> suggestions?
>
