### 2018-08-27 #671 @hooman650
Thank you, just felt like to leave a short note in regards to the research importance of BioSigKit. Please see below 35 "peer-reviewed" citations in regards to only "three" subroutines of BioSigKit:
![citations](https://user-images.githubusercontent.com/22506603/44679407-e304c100-aa08-11e8-8fd2-43367ea54670.PNG)

I have created BioSigKit based on numerous requests from fellow researchers that were simply asking me to put all of my highly cited work together in a single toolbox. Since many subroutines such as for instance "Activity detection with Hilbert Transform"([which happens to be the pick of the  week by mathworks](https://blogs.mathworks.com/pick/2014/05/23/automatic-activity-detection-using-hilbert-transform/)), "Activity detection based on processing 3 channel Accelerometer processing" and many more cannot simply be contributed to other existing projects, I hope that you agree that BioSigKit can be a helpful general purpose signal processing toolbox like [OSET](http://www.oset.ir/) which is not maintained anymore.

Thank you again,
Best
Hooman
