### 2018-09-10 #671 @Kevin-Mattheus-Moerman
@fernandoandreotti thanks again for your initial review. Could you please provide feedback on how the author has now dealt with your suggestions?
