### 2018-09-10 #671 @Kevin-Mattheus-Moerman
@marianux thanks for your reply and for your review work. I understand your concern in relation to this package being similar to previously published works. After considering your feedback I feel this work is sufficiently original and independent enough to warrant being a potential publication in JOSS. Given this, would you be able to provide feedback (and tick boxes at the top of this issue where appropriate) in relation to the other aspects of this work? Thanks.
