### 2018-09-29 d450a55
Update BioSigKit.m
