### 2018-10-10 #671 @Kevin-Mattheus-Moerman
@fernandoandreotti @marianux thanks for your review efforts so far. I should stress that novelty/independence in the sense that you are indicating is not a requirement for JOSS. Would you be able to look beyond this point and please comment on the other aspects of this work? Please also tick the boxes where appropriate.
