### 2018-10-11 #671 @cMadan
@Kevin-Mattheus-Moerman, I don't have time to conduct a thorough review of the functionality and documentation of the toolbox, but it does look like something that is well within the scope of JOSS, generally well-developed, and is sufficiently substantive to standalone as a toolbox.

This also means I disagree with some of the other reviewers that suggest this toolbox should be integrated into another open-source project.
