### 2018-10-17 #671 @Kevin-Mattheus-Moerman
@hooman650 please the final remaining issue I've just posted with some suggested minor edits to the paper (https://github.com/hooman650/BioSigKit/issues/6).

Once you've implemented those changes, please create a new release of your software and archive it in a service like [ZENODO](http://zenodo.org/). Once you've done that please list the DOI of the archived version here and also please mention the release/version number of the final reviewed and archived software. I'll then recommend that this work is accepted in JOSS.
