### 2018-10-17 #671 @hooman650
@Kevin-Mattheus-Moerman Thanks a lot for moving this forward. I have archived the final release on ZENODO.

[![DOI](https://zenodo.org/badge/128659224.svg)](https://zenodo.org/badge/latestdoi/128659224)

Version: V1.1

Thank you again.
