### 2018-10-17 #671 @hooman650
I also wanted to thank all of you guys on this thread! I definitely learned a lot about building and maintaining a clean repository! It would definitely help a lot for my future software submissions!

Cheers
