### 2018-10-18 #671 @Kevin-Mattheus-Moerman
@arfon We've had two completed and positive reviews (I've personally reviewed this submission as well). Some concerns were expressed by two initial reviewers (who have not completed the review process) who felt this work overlapped significantly with other projects. However myself and @cMadan conclude this work is sufficiently substantive to standalone as a toolbox. We therefore recommend acceptance of this work in JOSS. The author lists the updated version number, and the DOI of the archived version, above.
